backend/requirements.txt
# NovaTrend FastAPI 백엔드 의존성

# 핵심 프레임워크
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# HTTP 클라이언트 (뉴스 수집)
httpx==0.28.0
aiohttp==3.11.0

# 감성 분석 AI
transformers==4.47.0
torch==2.5.0
sentencepiece==0.2.0

# 데이터 처리
pydantic==2.10.0
feedparser==6.0.11

# Supabase 연동
supabase==2.11.0

# 스케줄러 (정기 수집)
apscheduler==3.10.4

# 테스트
pytest==8.3.0
pytest-asyncio==0.24.0
httpx[http2]==0.28.0
